Move `AffineQuantizedTensor` to torchao/dtypes
Summary:
att

Test Plan:
regression tests in test/quantization/test_quant_api.py

Reviewers:

Subscribers:

Tasks:

Tags:
